Function oak_functions_sdk::result_from_status

pub fn result_from_status<T>(status: i32, val: T) -> Result<T, OakStatus>

Convert a status returned from a host function call to a Result.

The status is interpreted as an int representing an OakStatus enum value.

If the status is [OakStatus::Ok] then returns the provided value as Result::Ok, otherwise returns the status as Result::Err.

Note that host function calls usually return an u32 because of limitations of the Wasm type system, so these values would usually be converted (via a cast) to i32 by callers.