Crate oak_functions_sdk

SDK functionality that provides idiomatic Rust wrappers around the underlying Oak Functions platform functionality.

Macros

log

Logs a debug message.

Functions

alloc

read_request

Reads and returns the user request.

result_from_status

Convert a status returned from a host function call to a Result.

storage_get_item

Looks up an item from the in-memory lookup store.

testing

Calls the testing extension with the given request. The response is directly passed on, as it is decoded by the caller.

write_log_message

Writes a debug log message.

write_response

Write the response.